+++ scripts/_helpers.py
@@ -1,5 +1,9 @@
-from pathlib import Path
+# SPDX-FileCopyrightText: Contributors to technology-data <https://github.com/pypsa/technology-data>
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
 import re
+from pathlib import Path
 
 
 class Dict(dict):
@@ -7,17 +11,15 @@ class Dict(dict):
     Dict is a subclass of dict, which allows you to get AND SET items in the
     dict using the attribute syntax!
 
-    Stripped down from addict https://github.com/mewwts/addict/ used in from pypsa.decriptor import Dict.
+    Stripped down from addict https://github.com/mewwts/addict/ used in from pypsa.descriptor import Dict.
     """
 
     def __setattr__(self, name, value):
         """
-        setattr is called when the syntax a.b = 2 is used to set a value.
+        Setattr is called when the syntax a.b = 2 is used to set a value.
         """
         if hasattr(Dict, name):
-            raise AttributeError(
-                "'Dict' object attribute " "'{0}' is read-only".format(name)
-            )
+            raise AttributeError("'Dict' object attribute " f"'{name}' is read-only")
         self[name] = value
 
     def __getattr__(self, item):