+++ scripts/retrieve_data_from_dea.py
@@ -1,5 +1,8 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
+# SPDX-FileCopyrightText: Contributors to technology-data <https://github.com/pypsa/technology-data>
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
+# coding: utf-8
 """
 Created on Mon May  4 18:48:11 2020
 
@@ -8,9 +11,10 @@ script retrieves technology data sheets from the danish energy agency
 @author: bw0928
 """
 
-import requests
-import urllib.request
 import time
+import urllib.request
+
+import requests
 from bs4 import BeautifulSoup
 
 # %%
@@ -24,31 +28,27 @@ response = requests.get(url)
 soup = BeautifulSoup(response.text, "html.parser")
 
 # %%
-search =  "https://ens.dk/en/our-services/projections-and-models/technology-data/"
-links = soup.findAll('a', {"href" : lambda href: href and search in href})
+search = "https://ens.dk/en/our-services/projections-and-models/technology-data/"
+links = soup.findAll("a", {"href": lambda href: href and search in href})
 
 for i in range(len(links)):
-
     one_a_tag = links[i]
     link_to_site = one_a_tag["href"]
     response2 = requests.get(link_to_site)
     soup2 = BeautifulSoup(response2.text, "html.parser")
-    data = soup2.findAll('a', {"href" : lambda href: href and ".xlsx" in href})
-    docu = soup2.findAll('a', {"href" : lambda href: href and ".pdf" in href})
+    data = soup2.findAll("a", {"href": lambda href: href and ".xlsx" in href})
+    docu = soup2.findAll("a", {"href": lambda href: href and ".pdf" in href})
 
     # get the data
     for j in range(len(data)):
-
         link_to_data = data[j]["href"]
         download_url = prefix + link_to_data
-        urllib.request.urlretrieve(download_url, path_out +
-                                   link_to_data.split("/")[-1])
+        urllib.request.urlretrieve(download_url, path_out + link_to_data.split("/")[-1])
         time.sleep(1)
 
     # get the documentation
     if docu:
         for j in range(len(docu)):
-
             link_to_docu = docu[j]["href"]
 
             if prefix not in link_to_docu:
@@ -56,11 +56,7 @@ for i in range(len(links)):
             else:
                 download_url = link_to_docu
 
-            urllib.request.urlretrieve(download_url, path_out + "/docu/" +
-                                       link_to_docu.split("/")[-1])
+            urllib.request.urlretrieve(
+                download_url, path_out + "/docu/" + link_to_docu.split("/")[-1]
+            )
             time.sleep(1)
-
-
-
-
-